+++ scripts/xyz_grid_on.py
@@ -236,13 +236,18 @@ class XYZGridScript(scripts_manager.Script):
                 valslist = [x for x in valslist_ext if x not in valslist]
             elif opt.type == str_permutations: # pylint: disable=comparison-with-callable
                 valslist = list(permutations(valslist))
-            valslist = [opt.type(x) for x in valslist]
+            elif opt.type == bool:
+                valslist = [str(x).strip().lower() in ['true', 't', 'yes', 'y', '1', 'on'] for x in valslist]
+            else:
+                valslist = [opt.type(x) for x in valslist]
             # Confirm options are valid before starting
             if opt.confirm:
                 opt.confirm(p, valslist)
             return valslist
 
         def parse_axis(x_type, x_values, x_values_dropdown):
+            if x_type is None:
+                x_type = 0
             x_opt = None
             if isinstance(x_type, str):
                 x_opt = [o for o in self.current_axis_options if o.label.lower() == x_type.lower()]